Changes to support clang

+++ include/revolution/types.h
@@ -14,11 +14,11 @@
 #include <decomp.h>
 #include <macros.h>
 
-typedef unsigned long long u64;
-typedef signed long long s64;
+typedef unsigned long u64;
+typedef signed long s64;
 
-typedef unsigned long u32;
-typedef signed long s32;
+typedef unsigned int u32;
+typedef signed int s32;
 
 typedef unsigned short u16;
 typedef signed short s16;
